# service/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clinic Antiviral Reporter",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/export-preview": {
      "get": {
        "summary": "Preview",
        "operationId": "getExportPreview",
        "parameters": [
          {
            "name": "dateFrom",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Datefrom"
            }
          },
          {
            "name": "dateTo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dateto"
            }
          },
          {
            "name": "manualSelection",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Manualselection"
            }
          },
          {
            "name": "selected",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Selected"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-session-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportPreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Review Selection",
        "operationId": "reviewExportSelection",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-session-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewSelection"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportPreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exports": {
      "post": {
        "summary": "Generate",
        "operationId": "createExport",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-session-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExport"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/SessionView"
                    },
                    {
                      "$ref": "#/components/schemas/PairingView"
                    },
                    {
                      "$ref": "#/components/schemas/DeviceView"
                    },
                    {
                      "$ref": "#/components/schemas/RefreshView"
                    },
                    {
                      "$ref": "#/components/schemas/CaseView"
                    },
                    {
                      "$ref": "#/components/schemas/BulkLotView"
                    },
                    {
                      "$ref": "#/components/schemas/ScanView"
                    },
                    {
                      "$ref": "#/components/schemas/MappingView"
                    }
                  ],
                  "title": "Response Createexport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exports/{export_id}/file": {
      "get": {
        "summary": "Download",
        "operationId": "downloadExport",
        "parameters": [
          {
            "name": "export_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Export Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-session-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BulkLotView": {
        "properties": {
          "appliedCases": {
            "items": {
              "$ref": "#/components/schemas/CaseView"
            },
            "type": "array",
            "title": "Appliedcases"
          }
        },
        "type": "object",
        "required": [
          "appliedCases"
        ],
        "title": "BulkLotView"
      },
...
      "CaseView": {
        "properties": {
          "caseId": {
            "type": "string",
            "title": "Caseid"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "chartNumber": {
            "type": "string",
            "title": "Chartnumber"
          },
          "patientName": {
            "type": "string",
            "title": "Patientname"
          },
          "birthDate": {
            "type": "string",
            "format": "date",
            "title": "Birthdate"
          },
          "physician": {
            "type": "string",
            "title": "Physician"
          },
          "reportingDate": {
            "type": "string",
            "format": "date",
            "title": "Reportingdate"
          },
          "sourceOrder": {
            "type": "string",
            "title": "Sourceorder"
          },
          "sourceQuantity": {
            "type": "integer",
            "title": "Sourcequantity"
          },
          "reportedQuantity": {
            "type": "integer",
            "title": "Reportedquantity"
          },
          "material": {
            "type": "string",
            "title": "Material"
          },
          "overdue": {
            "type": "boolean",
            "title": "Overdue"
          },
          "duplicateConcern": {
            "type": "boolean",
            "title": "Duplicateconcern"
          },
          "status": {
            "type": "string",
            "enum": [
              "awaiting_reason",
              "awaiting_reconciliation",
              "internally_complete",
              "excluded"
            ],
            "title": "Status",
            "default": "awaiting_reason"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "lots": {
            "items": {
              "$ref": "#/components/schemas/LotAllocation"
            },
            "type": "array",
            "title": "Lots"
          },
          "excluded": {
            "type": "boolean",
            "title": "Excluded",
            "default": false
          },
          "exclusionReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exclusionreason"
          },
          "sourceReviewRequired": {
            "type": "boolean",
            "title": "Sourcereviewrequired",
            "default": false
          },
          "sourceUnresolved": {
            "type": "boolean",
            "title": "Sourceunresolved",
            "default": false
          },
          "reportingSourceSnapshot": {
            "type": "integer",
            "title": "Reportingsourcesnapshot",
            "default": 0
          },
          "latestSourceSnapshot": {
            "type": "integer",
            "title": "Latestsourcesnapshot",
            "default": 0
          },
          "sourceDifferences": {
            "additionalProperties": {
              "additionalProperties": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "object"
            },
            "type": "object",
            "title": "Sourcedifferences"
          },
          "sourceTreatment": {
            "type": "string",
            "title": "Sourcetreatment",
            "default": "Y"
          },
          "internallyComplete": {
            "type": "boolean",
            "title": "Internallycomplete",
            "default": false
          },
          "exportEligible": {
            "type": "boolean",
            "const": false,
            "title": "Exporteligible",
            "default": false
          },
          "synthetic": {
            "type": "boolean",
            "const": true,
            "title": "Synthetic",
            "default": true
          },
          "liveIdentityVerified": {
            "type": "boolean",
            "const": false,
            "title": "Liveidentityverified",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "caseId",
          "revision",
          "chartNumber",
          "patientName",
          "birthDate",
          "physician",
          "reportingDate",
          "sourceOrder",
          "sourceQuantity",
          "reportedQuantity",
          "material",
          "overdue",
          "duplicateConcern"
        ],
        "title": "CaseView"
      },
      "CreateExport": {
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid",
            "title": "Requestid"
          },
          "expectedRevision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expectedrevision"
          },
          "cases": {
            "items": {
              "$ref": "#/components/schemas/BulkCaseRevision"
            },
            "type": "array",
            "minItems": 1,
            "title": "Cases"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requestId",
          "expectedRevision",
          "cases"
        ],
        "title": "CreateExport"
      },
...
      "DeviceView": {
        "properties": {
          "deviceId": {
            "type": "string",
            "title": "Deviceid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "capabilities": {
            "items": {
              "type": "string",
              "enum": [
                "admin",
                "physician",
                "reporting"
              ]
            },
            "type": "array",
            "title": "Capabilities"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "revoked": {
            "type": "boolean",
            "title": "Revoked"
          }
        },
        "type": "object",
        "required": [
          "deviceId",
          "name",
          "capabilities",
          "revision",
          "revoked"
        ],
        "title": "DeviceView"
      },
...
      "ExportCandidate": {
        "properties": {
          "case": {
            "$ref": "#/components/schemas/CaseView"
          },
          "defaultSelected": {
            "type": "boolean",
            "title": "Defaultselected"
          },
          "selected": {
            "type": "boolean",
            "title": "Selected"
          },
          "internalIssues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Internalissues"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "officialBlockers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Officialblockers"
          },
          "officiallyExportable": {
            "type": "boolean",
            "const": false,
            "title": "Officiallyexportable",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "case",
          "defaultSelected",
          "selected",
          "internalIssues",
          "warnings",
          "officialBlockers"
        ],
        "title": "ExportCandidate"
      },
...
      "ExportPreview": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExportCandidate"
            },
            "type": "array",
            "title": "Items"
          },
          "selectedCount": {
            "type": "integer",
            "title": "Selectedcount"
          },
          "internallyCompleteCount": {
            "type": "integer",
            "title": "Internallycompletecount"
          },
          "gates": {
            "items": {
              "$ref": "#/components/schemas/ExportGate"
            },
            "type": "array",
            "title": "Gates"
          },
          "operationalFlagEnabled": {
            "type": "boolean",
            "title": "Operationalflagenabled"
          },
          "productionExportEnabled": {
            "type": "boolean",
            "const": false,
            "title": "Productionexportenabled",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "items",
          "selectedCount",
          "internallyCompleteCount",
          "gates",
          "operationalFlagEnabled"
        ],
        "title": "ExportPreview"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MappingView": {
        "properties": {
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "goLiveAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goliveat"
          },
          "initialDateFrom": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Initialdatefrom"
          },
          "confirmedProfile": {
            "$ref": "#/components/schemas/MappingProfile",
            "default": {
              "internalCode": "ERA",
              "nhiCode": "A059653100",
              "materialValue": "DDMTR2018090002:易剋冒膠囊(顆)",
              "quantityRule": "integer_capsules"
            }
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/MappingVersion"
            },
            "type": "array",
            "title": "Versions"
          },
          "openGates": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Opengates",
            "default": [
              "live_his_contract",
              "official_export_contract",
              "synthetic_smis_import",
              "operator_authorization",
              "deployment_and_restore",
              "outage_recovery"
            ]
          },
          "productionExportEnabled": {
            "type": "boolean",
            "const": false,
            "title": "Productionexportenabled",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "revision",
          "goLiveAt",
          "initialDateFrom",
          "versions"
        ],
        "title": "MappingView"
      },
      "PairingView": {
        "properties": {
          "pairingCode": {
            "type": "string",
            "title": "Pairingcode"
          },
          "expiresAt": {
            "type": "number",
            "title": "Expiresat"
          }
        },
        "type": "object",
        "required": [
          "pairingCode",
          "expiresAt"
        ],
        "title": "PairingView"
      },
      "PreviewSelection": {
        "properties": {
          "dateFrom": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datefrom"
          },
          "dateTo": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dateto"
          },
          "selected": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Selected"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "selected"
        ],
        "title": "PreviewSelection"
      },
...
      "RefreshView": {
        "properties": {
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "created": {
            "type": "integer",
            "title": "Created"
          },
          "unchanged": {
            "type": "integer",
            "title": "Unchanged"
          },
          "changed": {
            "type": "integer",
            "title": "Changed",
            "default": 0
          },
          "quarantined": {
            "type": "integer",
            "title": "Quarantined",
            "default": 0
          },
          "synthetic": {
            "type": "boolean",
            "const": true,
            "title": "Synthetic",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "revision",
          "created",
          "unchanged"
        ],
        "title": "RefreshView"
      },
...
      "ScanView": {
        "properties": {
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "status": {
            "type": "string",
            "enum": [
              "disabled",
              "awaiting_configuration",
              "idle",
              "queued",
              "running",
              "succeeded",
              "partial",
              "failed",
              "interrupted"
            ],
            "title": "Status"
          },
          "jobId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jobid"
          },
          "dateFrom": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datefrom"
          },
          "dateTo": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dateto"
          },
          "startedAt": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat"
          },
          "finishedAt": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finishedat"
          },
          "lastSuccessAt": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastsuccessat"
          },
          "stale": {
            "type": "boolean",
            "title": "Stale",
            "default": true
          },
          "diagnostic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnostic"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Counts",
            "default": {}
          },
          "intervalSeconds": {
            "type": "integer",
            "title": "Intervalseconds"
          },
          "synthetic": {
            "type": "boolean",
            "const": true,
            "title": "Synthetic",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "revision",
          "enabled",
          "status",
          "intervalSeconds"
        ],
        "title": "ScanView"
      },
      "SessionView": {
        "properties": {
          "sessionId": {
            "type": "string",
            "title": "Sessionid"
          },
          "deviceId": {
            "type": "string",
            "title": "Deviceid"
          },
          "operator": {
            "type": "string",
            "title": "Operator"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Capabilities"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "expiresAt": {
            "type": "number",
            "title": "Expiresat"
          }
        },
        "type": "object",
        "required": [
          "sessionId",
          "deviceId",
          "operator",
          "capabilities",
          "revision",
          "expiresAt"
        ],
        "title": "SessionView"
      },
...
    }
  }
}